OOP Class for simple linear regression
by this class one can return mean and variance of feature and label as well as the covariance between them. Furthermore, printing the whole linear equation  or returning the slope and/or the intercept (All calculated by ordinary least square method)can be done by this Class. finally the predicted labels can be obtained.